refactor: domain패키지 생성, 해당되는 클래스 이동
