+++ src/main/java/nextstep/ladder/domain/Height.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 public class Height {
     public final int value;
