+++ src/main/java/nextstep/ladder/view/ResultView.java
@@ -1,6 +1,6 @@
 package nextstep.ladder.view;
 
-import nextstep.ladder.Ladder;
+import nextstep.ladder.domain.Ladder;
 
 public interface ResultView {
     void printLadder(Ladder ladder);
