+++ src/test/java/nextstep/ladder/LineTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder;
 
+import nextstep.ladder.domain.Line;
+import nextstep.ladder.domain.NumberOfParticipants;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
